PBR demo: expose internal resolution.
Former-commit-id: 48a57415916702986b0e05907bafd1f5673c80d7

+++ src/engine/Config.cpp
@@ -139,7 +139,7 @@ void RenderingConfig::processArguments(){
 		} else if(key == "fullscreen"){
 			fullscreen = true;
 		} else if(key == "internal-res" || key == "ivr"){
-			internalVerticalResolution = std::stof(values[0]);
+			internalVerticalResolution = std::stoi(values[0]);
 		} else if(key == "wxh"){
 			const unsigned int w = (unsigned int)std::stoi(values[0]);
 			const unsigned int h = (unsigned int)std::stoi(values[1]);
