+++ src/engine/Config.hpp
@@ -84,7 +84,7 @@ public:
 	
 	/// \brief Internal vertical rendering resolution.
 	/// \note The width should be computed based on the window aspect ratio.
-	float internalVerticalResolution = 720.0f;
+	int internalVerticalResolution = 720;
 	
 	/// Size of the window in raw pixels, updated at launch based on screen density.
 	glm::vec2 screenResolution = glm::vec2(800.0,600.0);
